requirements.txt
flask==3.1.0
python-dotenv==1.1.0
gunicorn==23.0.0

# LangChain
langchain==0.3.28
langchain-community==0.3.31
langchain-pinecone==0.2.8
langchain-google-genai==2.1.12
langchain-huggingface==0.1.2
sentence-transformers==3.3.1

# Vector DB
pinecone==7.3.0

# Database
supabase==2.11.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0

# Document loading
pypdf==5.6.1